Add files via upload
This code opens the .txt files we get from the Raman Spectroscopy application then stores it into a 3D array. The slices of the array are defined in a dictionary based off of the timestamp of the data. This program also gets the .csv files from the temperature controller and stores the timestamp and the temperature data of the .csv file in an array.